fix locations of tests, and the names of masterRoot tests
