+++ test/integration/prepareTransaction_test.js
